ZERO-24 [Refactor]: 타입 위치 변경
