+++ src/components/@shared/Dropdown.tsx
@@ -1,5 +1,26 @@
-import { DropdownProps, Option } from '@/src/types/dropdown';
-import React, { useCallback, useEffect, useRef, useState } from 'react';
+import React, {
+  ReactNode,
+  useCallback,
+  useEffect,
+  useRef,
+  useState,
+} from 'react';
+
+export interface Option {
+  label?: string;
+  component: ReactNode;
+}
+
+export interface DropdownProps {
+  triggerClass?: string; // 드롭다운 버튼에 추가할 클래스
+  optionsWrapClass?: string; // 옵션을 감싸는 div태그에 추가할 클래스
+  optionClass?: string; // 옵션에 추가할 클래스
+  triggerIcon?: ReactNode; // 드롭다운 버튼에 추가할 아이콘
+  initialOption?: Option;
+  options: Option[];
+  selected?: Option | null;
+  onSelect?: (option: Option) => void;
+}
 
 export default function Dropdown({
   initialOption,
